\name{concrete_slump}
\alias{concrete_slump}
\docType{data}
\title{Data Frame 'concrete_slump'}
\description{
This is the Concrete Slump Test data set containing various features of concrete mixtures and their slump test results.
}
\usage{data("concrete_slump")}
\format{
A data frame with
103
rows and
11
columns.
\itemize{
\item{\code{Cement}: Amount of cement in kg/m³.}
\item{\code{Slag}: Amount of blast furnace slag in kg/m³.}
\item{\code{Flyash}: Amount of fly ash in kg/m³.}
\item{\code{Water}: Amount of water in kg/m³.}
\item{\code{SP}: Amount of superplasticizer in kg/m³.}
\item{\code{Coarseagg}: Amount of coarse aggregate in kg/m³.}
\item{\code{Fineagg}: Amount of fine aggregate in kg/m³.}
\item{\code{SLUMP}: Slump of the concrete mixture in mm.}
\item{\code{FLOW}: Flow of the concrete mixture in mm.}
\item{\code{CompressiveStrength}: Compressive strength of the concrete in MPa.}
}
}
\examples{
data(concrete_slump)
}
\keyword{datasets}
