\name{cs_data_trial}
\alias{cs_data_trial}
\docType{data}
\title{
Cross-Sectional Dataset
}
\description{
Cross-Sectional Dataset
}
\usage{data("cs_data_trial")}
\format{
  A data frame with 1000 observations on the following 22 variables.
  \describe{
    \item{\code{name}}{a numeric vector}
    \item{\code{cons}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{u}}{a numeric vector}
    \item{\code{uz}}{a numeric vector}
    \item{\code{v}}{a numeric vector}
    \item{\code{u_t}}{a numeric vector}
    \item{\code{v_t}}{a numeric vector}
    \item{\code{u_c}}{a numeric vector}
    \item{\code{v_c}}{a numeric vector}
    \item{\code{u_e}}{a numeric vector}
    \item{\code{u_u}}{a numeric vector}
    \item{\code{y_pcs}}{a numeric vector}
    \item{\code{y_pcs_t}}{a numeric vector}
    \item{\code{y_pcs_e}}{a numeric vector}
    \item{\code{y_pcs_c}}{a numeric vector}
    \item{\code{y_pcs_u}}{a numeric vector}
    \item{\code{y_pcs_z}}{a numeric vector}
    \item{\code{y_pcs_w}}{a numeric vector}
    \item{\code{z}}{a numeric vector}
    \item{\code{con}}{a numeric vector}
  }
}
\details{
This is a trial run of data_gen_cs that allows the user to experiment with the main functions of the package, while knowing the true underlying parameters.  
}
\examples{
data(cs_data_trial)
}
\keyword{datasets}
