\name{data_gen_cs}
\alias{data_gen_cs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CS SF Models
}
\description{
SF models for CS 
}
\usage{
data_gen_cs(N, rand, sig_u, sig_v, cons, beta1, beta2, a, mu)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
Individuals
}
  \item{rand}{
random seed 
}
  \item{sig_u}{
sigma u 
}
  \item{sig_v}{
sigma v
}
  \item{cons}{
constant
}
  \item{beta1}{
beta 1 
}
  \item{beta2}{
beta 2
}
  \item{a}{
a from the THT 
}
\item{mu}{
mu from the Normal truncated normal
}
}
\details{
set the parameters above as desired for simulations
}
\author{
David Bernstein
}