\name{data_gen_cs}
\alias{data_gen_cs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CS SF Models
}
\description{
SF models for CS 
}
\usage{
data_gen_cs(N, rand, sig_u, sig_v, cons, beta1, beta2, a)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
HI
}
  \item{rand}{
HI
}
  \item{sig_u}{
HI
}
  \item{sig_v}{
HI
}
  \item{cons}{
HI
}
  \item{beta1}{
HI
}
  \item{beta2}{
HI
}
  \item{a}{
HI
}
}
\details{
HI
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
HI
}
\author{
David Bernstein
}
\note{
HI
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
HI
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

HI
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
