\name{protein}
\alias{protein}
\docType{data}
\title{Data Frame 'protein'}
\description{
This is the Protein Data Set containing various features related to protein structure and properties.
}
\usage{data("protein")}
\format{
A data frame with
45730
rows and
10
columns.
\itemize{
\item{\code{Feature1}: Description of Feature1.}
\item{\code{Feature2}: Description of Feature2.}
\item{\code{Feature3}: Description of Feature3.}
\item{\code{Feature4}: Description of Feature4.}
\item{\code{Feature5}: Description of Feature5.}
\item{\code{Feature6}: Description of Feature6.}
\item{\code{Feature7}: Description of Feature7.}
\item{\code{Feature8}: Description of Feature8.}
\item{\code{Feature9}: Description of Feature9.}
\item{\code{Feature10}: Description of Feature10.}
}
}
\examples{
data(protein)
}
\keyword{datasets}
