% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAPC.SFM.R
\name{SAPC.SFM}
\alias{SAPC.SFM}
\title{Stochastic Approximation Principal Component Analysis}
\usage{
SAPC.SFM(x, m, A, D, p)
}
\arguments{
\item{x}{The data used in the SAPC analysis.}

\item{m}{The number of common factors.}

\item{A}{The true factor loadings matrix.}

\item{D}{The true uniquenesses matrix.}

\item{p}{The number of variables.}
}
\value{
A list of metrics including:
\item{Asa}{Estimated factor loadings matrix obtained from the SAPC analysis.}
\item{Dsa}{Estimated uniquenesses vector obtained from the SAPC analysis.}
\item{MSESigmaA}{Mean squared error of the estimated factor loadings (Asa) compared to the true loadings (A).}
\item{MSESigmaD}{Mean squared error of the estimated uniquenesses (Dsa) compared to the true uniquenesses (D).}
\item{LSigmaA}{Loss metric for the estimated factor loadings (Asa), indicating the relative error compared to the true loadings (A).}
\item{LSigmaD}{Loss metric for the estimated uniquenesses (Dsa), indicating the relative error compared to the true uniquenesses (D).}
}
\description{
This function calculates several metrics for the SAPC method,
including the estimated factor loadings and uniquenesses, and various
error metrics comparing the estimated matrices with the true matrices.
}
\examples{

p = 10
m = 5
n = 2000
mu = t(matrix(rep(runif(p, 0, 100), n), p, n))
mu0 = as.matrix(runif(m, 0))
sigma0 = diag(runif(m, 1))
F = matrix(MASS::mvrnorm(n, mu0, sigma0), nrow = n)
A = matrix(runif(p * m, -1, 1), nrow = p)
xi = 5
omega = 2
alpha = 5
r <- sn::rsn(n * p, omega = omega, alpha = alpha) 
D0 = omega * diag(p)
D = diag(D0)
epsilon = matrix(r, nrow = n)
data = mu + F \%*\% t(A) + epsilon

result <- SAPC.SFM(data, m = m, A = A, D = D, p = p)
print(result)
}
