\name{sfm}
\alias{sfm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
CS SFM
}
\description{
CS SFM
}
\usage{
sfm(formula, model_name = c("START-NHN", "NHN", "NHN-MLQE", "NHN-PSI", "START-NHN_Z", "NHN_Z", "NHN_Z-MLQE", "NHN_Z-PSI", "START-THT", "THT", "THT-MLQE", "THT-PSI", "START-NE", "NE", "NE-MLQE", "NE-PSI", "START-CHC", "CHC", "CHC-MLQE", "CHC-PSI", "START-NU", "NU", "NU-MLQE", "NU-PSI"), data, maxit = 10000, maxit_2 = 1000, REPORT = 1, trace = 0, pgtol = 0, start_val = FALSE, inefdec = TRUE, gamma = FALSE, bob = TRUE, eta = 0.01, z_vars = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{formula}{
CS SFM
}
  \item{model_name}{
CS SFM
}
  \item{data}{
CS SFM
}
  \item{maxit}{
CS SFM
}
  \item{maxit_2}{
CS SFM
}
  \item{REPORT}{
CS SFM
}
  \item{trace}{
CS SFM
}
  \item{pgtol}{
CS SFM
}
  \item{start_val}{
CS SFM
}
  \item{inefdec}{
CS SFM
}
  \item{gamma}{
CS SFM
}
  \item{bob}{
CS SFM
}
  \item{eta}{
CS SFM
}
  \item{z_vars}{
CS SFM
}
}
\details{
CS SFM
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
CS SFM
}
\references{CS SFM
%% ~put references to the literature/web site here ~
}
\author{CS SFM
%%  ~~who you are~~
}
\note{CS SFM
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{CS SFM
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{CS SFM
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
  
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
