\name{yacht_hydrodynamics}
\alias{yacht_hydrodynamics}
\docType{data}
\title{Data Frame 'yacht_hydrodynamics'}
\description{
This is the Yacht Hydrodynamics data set containing various features of yacht design and their performance metrics.
}
\usage{data("yacht_hydrodynamics")}
\format{
A data frame with
364
rows and
7
columns.
\itemize{
\item{\code{LongitudinalPosition}: Longitudinal position of the center of buoyancy.}
\item{\code{PrismaticCoefficient}: Prismatic coefficient.}
\item{\code{LengthDisplacementRatio}: Length-displacement ratio.}
\item{\code{BeamDraughtRatio}: Beam-draught ratio.}
\item{\code{LengthBeamRatio}: Length-beam ratio.}
\item{\code{FroudeNumber}: Froude number.}
\item{\code{ResiduaryResistance}: Residuary resistance per unit weight of displacement.}
}
}
\examples{
data(yacht_hydrodynamics)
}
\keyword{datasets}
